+++ packages/components/pie-tag/src/index.ts
@@ -1,6 +1,7 @@
 import {
-    LitElement, html, unsafeCSS, nothing, type PropertyValues,
+    html, unsafeCSS, nothing, type PropertyValues,
 } from 'lit';
+import { PieElement } from '@justeattakeaway/pie-webc-core/src/internals/PieElement';
 import { property, queryAssignedElements } from 'lit/decorators.js';
 import { classMap, type ClassInfo } from 'lit/directives/class-map.js';
 import { validPropertyValues, defineCustomElement } from '@justeattakeaway/pie-webc-core';
@@ -23,7 +24,7 @@ const componentSelector = 'pie-tag';
  * @slot icon - The icon slot
  * @slot - Default slot
  */
-export class PieTag extends LitElement implements TagProps {
+export class PieTag extends PieElement implements TagProps {
     @property({ type: String })
     @validPropertyValues(componentSelector, variants, defaultProps.variant)
     public variant = defaultProps.variant;