+++ packages/tools/generator-pie-component/src/app/templates/declaration.d.ts
@@ -1,9 +1,0 @@
-declare module '*.scss' {
-    const content: Record<string, string>;
-    export default content;
-}
-
-declare module '*.scss?inline' {
-    const content: Record<string, string>;
-    export default content;
-}