+++ configs/pie-components-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @justeattakeaway/pie-components-config
 
+## 0.19.1
+
+### Patch Changes
+
+- [Changed] - use a centralised TS declaration file for common duplicated declares ([#2303](https://github.com/justeattakeaway/pie/pull/2303)) by [@jamieomaguire](https://github.com/jamieomaguire)
+
+## 0.19.0
+
+### Minor Changes
+
+- [Changed] - Add a v property to all components using vite to bake the version number into the code ([#2301](https://github.com/justeattakeaway/pie/pull/2301)) by [@jamieomaguire](https://github.com/jamieomaguire)
+
 ## 0.18.1
 
 ### Patch Changes