+++ packages/components/pie-textarea/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @justeattakeaway/pie-textarea
 
+## 0.15.1
+
+### Patch Changes
+
+- [Changed] - use a centralised TS declaration file for common duplicated declares ([#2303](https://github.com/justeattakeaway/pie/pull/2303)) by [@jamieomaguire](https://github.com/jamieomaguire)
+
+- Updated dependencies [[`edb0490`](https://github.com/justeattakeaway/pie/commit/edb049031b49a1d8a752c93db4cb164aa90eed3e)]:
+  - @justeattakeaway/pie-assistive-text@0.9.1
+  - @justeattakeaway/pie-webc-core@0.25.1
+
+## 0.15.0
+
+### Minor Changes
+
+- [Changed] - Add a v property to all components using vite to bake the version number into the code ([#2301](https://github.com/justeattakeaway/pie/pull/2301)) by [@jamieomaguire](https://github.com/jamieomaguire)
+
+### Patch Changes
+
+- Updated dependencies [[`87e902f`](https://github.com/justeattakeaway/pie/commit/87e902f4995af214770200e430eaec2bad7d21b3)]:
+  - @justeattakeaway/pie-assistive-text@0.9.0
+  - @justeattakeaway/pie-webc-core@0.25.0
+
 ## 0.14.1
 
 ### Patch Changes